% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aridanalysis.R
\name{arid_countreg}
\alias{arid_countreg}
\title{Function to create class object similar to sci-kit learn's object
structure for inferential purposes. Given a data frame, the response,
and certain specifications return a class object with a fit, predict, and
score functions as well as attributes obtained from the statistical analysis}
\usage{
arid_countreg(
  X,
  y,
  alpha = 0.05,
  fit_intercept = TRUE,
  verbose = FALSE,
  model = "additive",
  family = "poisson"
)
}
\arguments{
\item{X}{(data_frame): the input data frame with the explanatory variables to fit the model.}

\item{y}{(integer): an integer vector with the response to be fitted (only natural numbers).}

\item{alpha}{(double): a double vector of length 1 indicating the significance level (default: 0.05)}

\item{fit_intercept}{(logical): if the model should include the intercept (TRUE or FALSE). (default: FALSE)}

\item{verbose}{(logical): if results should include a written explanation (TRUE or FALSE). (default: FALSE)}

\item{model}{(character): type of model to be fitted, either "additive" or "interactive". (default: "additive")}

\item{family}{(character): distributional family to be used in generalized linear model. (default: "poisson")}
}
\value{
a class object with three methods and statistical attributes
}
\description{
Function to create class object similar to sci-kit learn's object
structure for inferential purposes. Given a data frame, the response,
and certain specifications return a class object with a fit, predict, and
score functions as well as attributes obtained from the statistical analysis
}
\examples{
X <- as.data.frame(matrix(rnorm(40 * 3), 40, 3))
y <- sample(c(1:60), 40, replace = TRUE)
arid_countreg(X,y,0.1)
}
