% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aridanalysis.R
\name{arid_eda}
\alias{arid_eda}
\title{Function to create summary statistics and basic EDA plots. Given a data frame,
this function outputs general exploratory analysis plots as well as basic
statistics summarizing trends in the features of the input data.}
\usage{
arid_eda(df, response, response_type = "numeric", features = c())
}
\arguments{
\item{response}{the column name of the response variable}

\item{response_type}{string indicating if response is 'categorical' or 'continuous'}

\item{features}{a list of explanatory variable column names}

\item{data_frame}{the input data frame to analyze}
}
\value{
a data frame with a list of features and their coefficients

a ggplot object containing the EDA
}
\description{
Function to create summary statistics and basic EDA plots. Given a data frame,
this function outputs general exploratory analysis plots as well as basic
statistics summarizing trends in the features of the input data.
}
