% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aridanalysis.R
\name{arid_linreg}
\alias{arid_linreg}
\title{Function that builds an arid_linreg class model object that provides sklearn
linear regression interface functionality and attributes. The arid_linreg
function instantiates a linear regression model type based on the input
specifications and provides methods to fit/predict/score the results and
retrieve the calculated sklearn coefficients.}
\usage{
arid_linreg(regularization = NULL, lambda = NULL)
}
\arguments{
\item{regularization}{(character): A string defining NULL,'L1','L2', or 'L1L2'
linear coefficient regularization (default: NULL)}

\item{lambda}{(double):the numeric regularization strength value}
}
\value{
arid_linreg: an arid_linreg class linear regression model object
}
\description{
Function that builds an arid_linreg class model object that provides sklearn
linear regression interface functionality and attributes. The arid_linreg
function instantiates a linear regression model type based on the input
specifications and provides methods to fit/predict/score the results and
retrieve the calculated sklearn coefficients.
}
\examples{
model <- arid_linreg()
model$fit(matrix(1:12, nrow = 3, ncol = 3), as.matrix(c(1,2,0)))
model$predict(t(as.matrix(c(1,2,2))))
model$score()
}
