% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aridanalysis.R
\name{arid_logreg}
\alias{arid_logreg}
\title{Given a matrix X of explanatory variables, a response y numeric variable,
this function fits either a 'binomial' or 'multinomial' logistic regression model
and returns a class object similar to sci-kit learn's object}
\usage{
arid_logreg(X, y, regularization = NULL, lambda = NULL)
}
\arguments{
\item{X}{(data frame): the explanatory variables matrix}

\item{y}{(numeric): the response variable numeric vector}

\item{regularization}{(character): what level of regularization to use in the model (default NULL)}

\item{lambda}{(double): the regularization strength parameter to use (default: NULL)}
}
\value{
arid_logreg: a class object after fitting a 'binomial' or 'multinomial' logistic regression model
}
\description{
Given a matrix X of explanatory variables, a response y numeric variable,
this function fits either a 'binomial' or 'multinomial' logistic regression model
and returns a class object similar to sci-kit learn's object
}
\examples{
X <- matrix(rnorm(40 * 3), 40, 3)
y <- sample(c(0,1), 40, replace = TRUE)
arid_logreg(X, y)
}
