% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aridanalysis.R
\name{aRid_linreg}
\alias{aRid_linreg}
\title{Function that builds an aRid_linreg class model object that provides sklearn
linear regression interface functionality and attributes. The aRid_linreg
function instantiates a linear regression model type based on the input
specifications and provides methods to fit/predict/score the results and
retrieve sklearn attributes.}
\usage{
aRid_linreg(regularization = NULL, lambda = NULL)
}
\arguments{
\item{regularization}{A string defining NULL,'L1','L2', or 'L1L2'
linear coefficient regularization}

\item{lambda}{the numeric regularization strength value}
}
\value{
an aRid_linreg class linear regression model object
}
\description{
Function that builds an aRid_linreg class model object that provides sklearn
linear regression interface functionality and attributes. The aRid_linreg
function instantiates a linear regression model type based on the input
specifications and provides methods to fit/predict/score the results and
retrieve sklearn attributes.
}
\examples{
model <- aRid_linreg()
model$fit(matrix(1:12, nrow = 3, ncol = 3), as.matrix(c(1,2,0)))
model$predict(t(as.matrix(c(1,2,2))))
model$score()
}
